backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
bcrypt==4.1.1
python-jose[cryptography]==3.3.0
pyodbc==5.2.0
python-dotenv==1.0.0
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
openai>=1.12.0